Improved workflow through generalization and more
The JIRAIssue has become more general and useful. There are several new command
line options for easy configurability and ticket creation.

DAN tickets are now supported for easy testing (currently they do not have the
same structure as CS tickets which makes CS testing difficult). Template files
now support terminating a line with '\' so that they may be formatted nicely in
in the template file but appear as fluid single lines in the ticket.