Some reorganization and general betterment
Reorganize and provide somewhat better comments in supportissue.py. Got rid of
hypothesis crap in lastXGenOnlyComment determination. Fixed bug in
earliestOfLastCustomerComments method.

Change getTransitionId in jirapp to query JIRA each time for the ticket status
before performing the transition as a ticket may undergo several state-changing
actions between the time we first queried it in our local db and the time of
the transition.