Pass find query to getListOfTickets
The getListOfTickets method now takes an optional match dict. This is used by
the API to only act on active tickets, i.e not done, in the case of /queue, and
only active tickets in the given workflow in the case of /workflow.

By default getListOfTickets does not use a match and hence returns ALL tickets.
The /queue call therefore gets ALL tickets, even done ones. At some point we
may want to modify this endpoint or add another so that it will only return a
list of active tickets if requested.