Replace OptionParser with ArgumentParser
Replace OptionParser (deprecated) with ArgumentParser. Add new CLI arguments
including --dry-run, --jira-username, --jira-password, --limit.

A config file is no longer required as everything is configurable on the CLI.

However, one may still specify a config file. Each line of the config file must
either start with a '#' to signify a comment, or start with a --CLI-argument
followed by a value if the same CLI argument would expect a value.