New cli/daemon mode, new logging, optional REST
There are now two modes of working with karakuri: via the command line and as
a daemon.

See the various -h for cli options.

The daemon can be started with a RESTful interface (--rest). Otherwise it loops
and does not currently do anything useful.

Logging has been cleaned up a lot, there is one annoying thing still where
things are logged twice because messages are sent both to my filehandler and
to stderr/stdout, which is hella annoying.